Add files via upload
These files are matrices that represent pairwise distance values between species of moth, butterfly and bumblebee traits, phylogeny and population size. 
"all_pop.rds" is a matrix of population asynchrony between every pair of species used in the study. 
"bee_phylo.rds" is a matrix of phylogenetic distance between every pair of bumblebee species used in the study.
"bee_pop.rds" is a matrix of population asynchrony between every pair of bumblebee species used in the study. 
"bee_trait_dist.rds" is a matrix of Gower distances between traits of every pair of bumblebee species used in the study. 
"but_phylo.rds" is a matrix of phylogenetic distance between every pair of butterfly species used in the study.
"but_pop.rds" is a matrix of population asynchrony between every pair of butterfly species used in the study. 
"but_trait_dist.rds" is a matrix of Gower distances between traits of every pair of butterfly species used in the study.
"mb_phylo.rds" is a matrix of phylogenetic distance between every pair of moth and butterfly species used in the study.
"mb_pop.rds" is a matrix of population asynchrony between every pair of moth and butterfly species used in the study. 
"lep_trait_dist.rds" is a matrix of Gower distances between traits of every pair of moth and butterfly species used in the study.
"moth_phylo.rds" is a matrix of phylogenetic distance between every pair of moth species used in the study.
"moth_pop.rds" is a matrix of population asynchrony between every pair of moth species used in the study. 
"moth_trait_dist.rds" is a matrix of Gower distances between traits of every pair of moth species used in the study.